+++ test/entrypoint/srv_bucket_id_pk/storage.lua
@@ -1,0 +1,42 @@
+local helper = require('test.helper')
+
+return {
+    init = helper.wrap_schema_init(function()
+        local engine = os.getenv('ENGINE') or 'memtx'
+        local customers_space = box.schema.space.create('customers', {
+            format = {
+                {name = 'id', type = 'unsigned'},
+                {name = 'bucket_id', type = 'unsigned'},
+                {name = 'name', type = 'string'},
+                {name = 'age', type = 'number'},
+            },
+            if_not_exists = true,
+            engine = engine,
+        })
+        customers_space:create_index('bucket_id', {
+            parts = { {field = 'bucket_id'}, {field = 'id'} },
+            if_not_exists = true,
+        })
+
+        -- https://github.com/tarantool/migrations/blob/a7c31a17f6ac02d4498b4203c23e495856861444/migrator/utils.lua#L35-L53
+        if box.space._ddl_sharding_key == nil then
+            local sharding_space = box.schema.space.create('_ddl_sharding_key', {
+                format = {
+                    {name = 'space_name', type = 'string', is_nullable = false},
+                    {name = 'sharding_key', type = 'array', is_nullable = false}
+                },
+                if_not_exists = true,
+            })
+            sharding_space:create_index(
+                'space_name', {
+                    type = 'TREE',
+                    unique = true,
+                    parts = {{'space_name', 'string', is_nullable = false}},
+                    if_not_exists = true,
+                }
+            )
+        end
+        box.space._ddl_sharding_key:replace{'customers', {'id'}}
+    end),
+    wait_until_ready = helper.wait_schema_init,
+}